Signin UI and app.js changes
