+++ src/App.css
@@ -1,1 +1,4 @@
-
+@import url('https://fonts.googleapis.com/css2?family=Nanum+Gothic:wght@400;800&display=swap');
+* {
+	font-family: 'Nanum Gothic', sans-serif;
+}
